fix: scriptLoader => index. Add proper index tags
