+++ src/classes/redis-connection.ts
@@ -5,11 +5,13 @@ import { Cluster, Redis } from 'ioredis';
 // @ts-ignore
 import { CONNECTION_CLOSED_ERROR_MSG } from 'ioredis/built/utils';
 import * as semver from 'semver';
-import { load, loadIncludes } from '../commands';
-import { ConnectionOptions, RedisOptions, RedisClient } from '../interfaces';
+import { load } from '../commands';
+import { ConnectionOptions, RedisOptions } from '../interfaces';
 import { isRedisInstance, isNotConnectionError } from '../utils';
 
 import * as path from 'path';
+
+export type RedisClient = Redis | Cluster;
 
 const overrideMessage = [
   'BullMQ: WARNING! Your redis options maxRetriesPerRequest must be null and enableReadyCheck false',
@@ -42,7 +44,7 @@ export class RedisConnection extends EventEmitter {
       this.opts = {
         port: 6379,
         host: '127.0.0.1',
-        retryStrategy: function (times: number) {
+        retryStrategy: function(times: number) {
           return Math.min(Math.exp(times), 20000);
         },
         ...opts,
@@ -121,10 +123,6 @@ export class RedisConnection extends EventEmitter {
 
   protected loadCommands(): Promise<void> {
     return load(this._client, path.join(__dirname, '../commands'));
-  }
-
-  protected loadIncludes(): Promise<Record<string, string>> {
-    return loadIncludes(path.join(__dirname, '../commands'));
   }
 
   private async init() {
