+++ src/commands/moveToFinished-8.lua
@@ -42,8 +42,8 @@
 local rcall = redis.call
 
 -- Includes
-<%= updateParentDepsIfNeeded %>
-<%= destructureJobKey %>
+--- @include "includes/updateParentDepsIfNeeded"
+--- @include "includes/destructureJobKey"
 
 local jobIdKey = KEYS[3]
 if rcall("EXISTS",jobIdKey) == 1 then -- // Make sure job exists
@@ -70,8 +70,8 @@ if rcall("EXISTS",jobIdKey) == 1 then -- // Make sure job exists
     if(numRemovedElements < 1) then
       return -3
     end
-    
-    -- If job has a parent we need to 
+
+    -- If job has a parent we need to
     -- 1) remove this job id from parents dependencies
     -- 2) move the job Id to parent "processed" set
     -- 3) push the results into parent "results" list
@@ -83,15 +83,15 @@ if rcall("EXISTS",jobIdKey) == 1 then -- // Make sure job exists
       parentId = getJobIdFromKey(ARGV[14])
       parentQueueKey = getJobKeyPrefix(ARGV[14], ":" .. parentId)
     end
-    if parentId ~= "" and ARGV[5] == "completed" then 
+    if parentId ~= "" and ARGV[5] == "completed" then
         local parentKey =  parentQueueKey .. ":" .. parentId
         local dependenciesSet = parentKey .. ":dependencies"
         local result = rcall("SREM", dependenciesSet, jobIdKey)
-        if result == 1 then 
+        if result == 1 then
           updateParentDepsIfNeeded(parentKey, parentQueueKey, dependenciesSet, parentId, jobIdKey, ARGV[4])
         end
     end
-    
+
     -- Remove job?
     local removeJobs = tonumber(ARGV[6])
     if removeJobs ~= 1 then
