+++ src/commands/removeJob-1.lua
@@ -13,7 +13,7 @@
 local rcall = redis.call
 
 -- Includes
-<%= destructureJobKey %>
+--- @include "includes/destructureJobKey"
 
 -- recursively check if there are no locks on the
 -- jobs to be removed.
@@ -45,13 +45,13 @@ local function removeJob( prefix, jobId)
     local jobKey = prefix .. jobId;
 
     -- Check if this job has a parent. If so we will just remove it from
-    -- the parent child list, but if it is the last child we should move the parent to "wait/paused" 
+    -- the parent child list, but if it is the last child we should move the parent to "wait/paused"
     -- which requires code from "moveToFinished"
     local parentKey = rcall("HGET", jobKey, "parentKey")
     if( (type(parentKey) == "string") and parentKey ~= "" and (rcall("EXISTS", parentKey) == 1)) then
         local parentDependenciesKey = parentKey .. ":dependencies"
         local result = rcall("SREM", parentDependenciesKey, jobKey)
-        if rcall("SCARD", parentDependenciesKey) == 0 then 
+        if rcall("SCARD", parentDependenciesKey) == 0 then
             local parentId = getJobIdFromKey(parentKey)
             local parentPrefix = getJobKeyPrefix(parentKey, parentId)
